requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
elastic-transport==8.12.0
elasticsearch==8.12.0
idna==3.6
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.0
